Update some rules. (#428)
* Sofire

There are multiple applications that have libhs.so, which is not Hyperscan, but Sofire, which can be misleading. The existing Hyperscan rule description has some errors and there are more applications using Sofire, so it is more appropriate to change it to Sofire.

* Sofire

* Sofire

* Sofire

* MACE

libmotion_photo_mace.so is Xiaomi's private library, which exists only in Bokeh. Only libmace.so should be kept.